+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add drug stock option in progress
 - Show custom error screen for drug stock web view when there is no internet connection
 - Add feature flag for village type ahead search
+- Remove date of birth `DateTimeFormatter` as constructor param from `PatientRepository`
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding